Move HTTP code to a sub-package.
