+++ src/main/java/uk/co/amrc/factoryplus/FPDiscovery.java
@@ -13,6 +13,8 @@ import java.util.UUID;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import uk.co.amrc.factoryplus.http.*;
+
 public class FPDiscovery {
     private static final Logger log = LoggerFactory.getLogger(FPHttpClient.class);
